+++ src/rgr.rs
@@ -11,6 +11,7 @@ fn main() -> std::io::Result<()> {
         .propagate_version(true)
         .arg_required_else_help(true)
         .subcommand(cmd_rgr::count::make_subcommand())
+        .subcommand(cmd_rgr::field::make_subcommand())
         .subcommand(cmd_rgr::merge::make_subcommand())
         .subcommand(cmd_rgr::replace::make_subcommand())
         .subcommand(cmd_rgr::runlist::make_subcommand());
@@ -18,6 +19,7 @@ fn main() -> std::io::Result<()> {
     // Check which subcomamnd the user ran...
     match app.get_matches().subcommand() {
         Some(("count", sub_matches)) => cmd_rgr::count::execute(sub_matches),
+        Some(("field", sub_matches)) => cmd_rgr::field::execute(sub_matches),
         Some(("merge", sub_matches)) => cmd_rgr::merge::execute(sub_matches),
         Some(("replace", sub_matches)) => cmd_rgr::replace::execute(sub_matches),
         Some(("runlist", sub_matches)) => cmd_rgr::runlist::execute(sub_matches),
@@ -28,5 +30,4 @@ fn main() -> std::io::Result<()> {
     Ok(())
 }
 
-// TODO: `rgr field`
 // TODO: `rgr annotate`